Put completion tests for different shells in submodules
